Begin with permissions

+++ users/api.py
@@ -2,6 +2,8 @@ from django.contrib.auth.models import User
 from rest_framework.status import HTTP_201_CREATED, HTTP_400_BAD_REQUEST, HTTP_202_ACCEPTED, HTTP_204_NO_CONTENT
 from rest_framework.views import APIView
 from rest_framework.response import Response
+
+from users.permissions import UserPermission
 from users.serializers import UserSerializer, UserListSerializer
 from django.shortcuts import get_object_or_404
 
@@ -10,6 +12,9 @@ class UserListAPI(APIView):
     """
     Endpoint de listado de usuarios
     """
+
+    permission_classes = (UserPermission,)
+
     def get(self, request):
         users = User.objects.all()
         serializer = UserListSerializer(users, many=True)
@@ -27,14 +32,17 @@ class UserDetailAPI(APIView):
     """
     Endpoint of user detail
     """
+    permission_classes = (UserPermission,)
 
     def get(self, request, pk):
         user = get_object_or_404(User, pk=pk) #si no encuentra ningún usuario, devuelve una respuesta 404
+        self.check_object_permissions(request, user)  # llama a has_object_permission
         serializer = UserSerializer(user)
         return Response(serializer.data)
 
     def put(self, request, pk):
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)  # llama a has_object_permission
         serializer = UserSerializer(user, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -44,5 +52,6 @@ class UserDetailAPI(APIView):
 
     def delete(self, request, pk):
         user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)  # llama a has_object_permission
         user.delete()
         return Response(status=HTTP_204_NO_CONTENT)
